Extended xml database loading and tests.
Added pointer to xmldb structure in runtime context. All loaded
databases are enlisted in an APR hash structure inside the xmldb
structure. Upgraded return status codes and read in case there are XML
structure errors. It is now required at least one database to be loaded
for the xmldb_init() method to be successful.